Fixed bug in python implementation, finished algorithm for javascript
I was turning the initial pool into a set, which was both unnecessary,
and scrambling the order of the list. This had the side-effect of making
the program guesses a bit irregular, and de-synchronized from the
javascript implementation, which doesn't have a set type. I switched it
to a list, and oddly, the guesses seem much more patterned. It'll be
interesting to compare the average number of guesses for both.